+++ openmetadata-ui/src/main/resources/ui/src/components/common/TabsPane/TabsPane.tsx
@@ -12,7 +12,7 @@
  */
 
 import classNames from 'classnames';
-import { isNil, lowerCase } from 'lodash';
+import { camelCase, isNil } from 'lodash';
 import React from 'react';
 import { TITLE_FOR_NON_OWNER_ACTION } from '../../../constants/constants';
 import { getCountBadge } from '../../../utils/CommonUtils';
@@ -58,7 +58,7 @@ const TabsPane = ({ activeTab, setActiveTab, tabs, className = '' }: Props) => {
                 <button
                   className={getTabClasses(tab.position, activeTab)}
                   data-testid="tab"
-                  id={lowerCase(tab.name)}
+                  id={camelCase(tab.name)}
                   onClick={() => setActiveTab?.(tab.position)}>
                   <SVGIcons
                     alt={tab.icon.alt}
@@ -77,7 +77,7 @@ const TabsPane = ({ activeTab, setActiveTab, tabs, className = '' }: Props) => {
               <button
                 className={getTabClasses(tab.position, activeTab)}
                 data-testid="tab"
-                id={lowerCase(tab.name)}
+                id={camelCase(tab.name)}
                 key={tab.position}
                 onClick={() => setActiveTab?.(tab.position)}>
                 <SVGIcons
