+++ openmetadata-ui/src/main/resources/ui/src/interface/types.d.ts
@@ -422,4 +422,13 @@ declare module 'Models' {
     openInNewTab?: boolean;
     showLabel?: boolean;
   };
+
+  export type TourSteps = {
+    content?: string | React.ReactNode;
+    actionType?: string;
+    position?: string | number[];
+    selector?: string;
+    userTypeText?: string;
+    waitTimer?: number;
+  };
 }
