+++ openmetadata-ui/src/main/resources/ui/src/pages/tour-page/TourPage.component.tsx
@@ -1,6 +1,7 @@
 import { observer } from 'mobx-react';
 import { LeafNodes, SearchResponse } from 'Models';
 import React, { useEffect, useState } from 'react';
+import { useLocation } from 'react-router-dom';
 import AppState from '../../AppState';
 import DatasetDetails from '../../components/DatasetDetails/DatasetDetails.component';
 import { DatasetOwner } from '../../components/DatasetDetails/DatasetDetails.interface';
@@ -8,7 +9,9 @@ import Explore from '../../components/Explore/Explore.component';
 import { ExploreSearchData } from '../../components/Explore/explore.interface';
 import MyData from '../../components/MyData/MyData.component';
 import { MyDataProps } from '../../components/MyData/MyData.interface';
+import NavBar from '../../components/nav-bar/NavBar';
 import Tour from '../../components/tour/Tour';
+import { ROUTES, TOUR_SEARCH_TERM } from '../../constants/constants';
 import {
   mockDatasetData,
   mockFeedData,
@@ -22,6 +25,7 @@ import {
 } from '../../generated/entity/data/table';
 import { TagLabel } from '../../generated/type/tagLabel';
 import { useTour } from '../../hooks/useTour';
+import { getSteps } from '../../utils/TourUtils';
 
 const mockData = {
   data: { hits: { hits: [] } },
@@ -36,6 +40,7 @@ const exploreCount = {
 };
 
 const TourPage = () => {
+  const location = useLocation();
   const { handleIsTourOpen } = useTour();
   const [currentPage, setCurrentPage] = useState<CurrentTourPageType>(
     AppState.currentTourPage
@@ -47,9 +52,27 @@ const TourPage = () => {
     AppState.activeTabforTourDatasetPage
   );
   const [explorePageCounts, setExplorePageCounts] = useState(exploreCount);
+  const [searchValue, setSearchValue] = useState('');
 
   const handleCountChange = () => {
     setExplorePageCounts(exploreCount);
+  };
+
+  const clearSearchTerm = () => {
+    setSearchValue('');
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    if (e.key === 'Enter') {
+      if (location.pathname.includes(ROUTES.TOUR)) {
+        if (searchValue === TOUR_SEARCH_TERM) {
+          AppState.currentTourPage = CurrentTourPageType.EXPLORE_PAGE;
+          clearSearchTerm();
+        }
+
+        return;
+      }
+    }
   };
 
   useEffect(() => {
@@ -164,7 +187,21 @@ const TourPage = () => {
 
   return (
     <div>
-      <Tour />
+      <NavBar
+        isTourRoute
+        handleFeatureModal={handleCountChange}
+        handleKeyDown={handleKeyDown}
+        handleSearchBoxOpen={handleCountChange}
+        handleSearchChange={(value) => setSearchValue(value)}
+        isFeatureModalOpen={false}
+        isSearchBoxOpen={false}
+        pathname={location.pathname}
+        profileDropdown={[]}
+        searchValue={searchValue}
+        settingDropdown={[]}
+        supportDropdown={[]}
+      />
+      <Tour steps={getSteps(TOUR_SEARCH_TERM, clearSearchTerm)} />
       {getCurrentPage(currentPage)}
     </div>
   );
